fix price + inventory

+++ src/main/java/Beauty_ECatalog/Beauty_ECatalog/domain/response/ResImportTicketDetail.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 public class ResImportTicketDetail {
     private String supplier;
     private Instant date;
-    private long total;
+    private double total;
     private boolean status;
     private List<ProductInTicket> listProducts;
